feat: use market data from bridged assets if base asset data is unavailable
